refactoring todoform
addTodo 関数で this.state.todos.length + 1 のように，最後に + 1 を付け加えるとエラーが出る
謎

+1 をつけると id が連番にならなくなり (id を 0 から始めているからというのもある) それが原因 ?
=> react は id は連番じゃないといけないの ?

react はじゃなくて javascript の map とか iteration で変になっている可能性も
