+++ src/components/Interfaces.ts
@@ -2,5 +2,9 @@ export interface Item {
   id: number;
   text: string;
   status: string;
-  completed: boolean;
+}
+
+export interface AddTodoProps {
+  text: string;
+  status: string;
 }
